openshift-analyzer/backend/requirements.txt
fastapi>=0.95.1,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<2.0.0
pyyaml>=6.0,<7.0
kubernetes>=26.1.0,<27.0.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.6,<0.1.0
httpx>=0.24.0,<0.25.0
jinja2>=3.1.2,<3.2.0
asyncio>=3.4.3,<3.5.0
python-ldap>=3.4.3,<3.5.0
aiofiles>=23.1.0,<24.0.0
prometheus-client>=0.16.0,<0.17.0
PyJWT>=2.6.0,<2.7.0
requests>=2.29.0,<2.30.0
jsonschema>=4.17.3,<4.18.0 